% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.clean_chr_name}
\alias{.clean_chr_name}
\title{Clean chromosome names for display}
\usage{
.clean_chr_name(x)
}
\arguments{
\item{x}{Character vector of chromosome names.}
}
\value{
Character vector with "chr" replaced by "Chr ".
}
\description{
Clean chromosome names for display
}
\keyword{internal}
