% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.elegans_chr_order}
\alias{.elegans_chr_order}
\title{Default C. elegans chromosome order}
\usage{
.elegans_chr_order()
}
\value{
Character vector of length 6.
}
\description{
Default C. elegans chromosome order
}
\keyword{internal}
