% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ercantools-package.R
\docType{package}
\name{ercantools-package}
\alias{ercantools}
\alias{ercantools-package}
\title{ercantools: Visualization and Analysis Tools for ChIP-seq Differential Binding Data}
\description{
\code{ercantools} provides a suite of plotting and statistical analysis functions
for ChIP-seq differential binding data, designed to work directly with
\href{https://bioconductor.org/packages/DiffBind/}{DiffBind} output.
\subsection{Function families}{
\itemize{
\item \strong{Violin plots}: \code{\link[=violin_log2FC]{violin_log2FC()}}, \code{\link[=violin_counts]{violin_counts()}},
\code{\link[=violin_log2FC_all_chr]{violin_log2FC_all_chr()}}, \code{\link[=violin_counts_all_chr]{violin_counts_all_chr()}}
\item \strong{T-test scatter plots}: \code{\link[=ttest_scatter]{ttest_scatter()}}
\item \strong{DiffBind diagnostic plots}: \code{\link[=plot_md]{plot_md()}}, \code{\link[=plot_distributions]{plot_distributions()}}
\item \strong{Peak utilities}: \code{\link[=find_overlap_peaks]{find_overlap_peaks()}}, \code{\link[=upset_peaks]{upset_peaks()}}
}
}

\subsection{Organism defaults}{

Functions are designed for \emph{C. elegans} (chrI-V + chrX) by default,
but all chromosome parameters are fully configurable for other organisms.
}

\subsection{Example data}{

A synthetic dataset mimicking DiffBind output is included:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data(example_peaks)
}\if{html}{\out{</div>}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ercanlab/ercantools}
  \item \url{https://ercanlab.github.io/ercantools}
  \item Report bugs at \url{https://github.com/ercanlab/ercantools/issues}
}

}
\author{
\strong{Maintainer}: Daniel Garbozo \email{dg4739@nyu.edu}

Other contributors:
\itemize{
  \item Sevinc Ercan \email{sercan@nyu.edu} [contributor]
}

}
\keyword{internal}
